+++ src/core/plugins/core-commands/server/commands/moderator/index.ts
@@ -1,6 +1,5 @@
 import './currency';
 import './door';
-import './garage';
 import './interior';
 import './noclip';
 import './player';