+++ src/core/server/api/consts/constSystem.ts
@@ -1,11 +1,12 @@
 import { MessengerSystem } from '@AthenaServer/systems/messenger';
-import { AgendaSystem } from '../../systems/agenda';
-import { Identifier } from '../../systems/identifier';
-import { ItemFactory } from '../../systems/item';
-import { ItemEffects } from '../../systems/itemEffects';
-import { addJobCheck, cloneObjective, getPlayerJob, Job } from '../../systems/job';
-import { PluginSystem } from '../../systems/plugins';
-import { SoundSystem } from '../../systems/sound';
+import { PermissionSystem } from '@AthenaServer/systems/permission';
+import { AgendaSystem } from '@AthenaServer/systems/agenda';
+import { Identifier } from '@AthenaServer/systems/identifier';
+import { ItemFactory } from '@AthenaServer/systems/item';
+import { ItemEffects } from '@AthenaServer/systems/itemEffects';
+import { addJobCheck, cloneObjective, getPlayerJob, Job } from '@AthenaServer/systems/job';
+import { PluginSystem } from '@AthenaServer/systems/plugins';
+import { SoundSystem } from '@AthenaServer/systems/sound';
 import { defaultSystemsConst } from './constDefaultSystems';
 
 export const systemConst = {
@@ -21,6 +22,7 @@ export const systemConst = {
         instance: Job,
     },
     messenger: MessengerSystem,
+    permission: PermissionSystem,
     plugins: PluginSystem,
     sound: SoundSystem,
 };